### 2016-06-16 #53 @sckott
mregions: Marine Regions Data from Marineregions.org
* 1. What does this package do? (explain in 50 words or less)
  `mregions` is a wrapper to REST API methods from http://www.marineregions.org/ - with purpose of fetching marine regions in geojson/shp/wkt format for use downstream.
* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: mregions
Title: Marine Regions Data from Marineregions.org
Description: Tools to get marine regions data from Marineregions.org.
    Includes tools to get region metadata, as well as data in GeoJSON
    format, as well as Shape files.
Version: 0.0.9.9400
    person("Scott", "Chamberlain", role = c("aut", "cre"), email = "myrmecocystus@gmail.com"),
    person("Pieter", "Provoost", role = "aut")
    )
    jsonlite (>= 0.9.20),
    xml2,
    wellknown,
    rappdirs
    geojsonio,
    rgdal,
    rgeos,
    knitr
Enhances: leaflet
```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/ropenscilabs/mregions
* 4. What data source(s) does it work with (if applicable)?
REST API from http://www.marineregions.org/ and a geoserver API from a related service http://geo.vliz.be/geoserver/ows
* 5. Who is the target audience?
Biologists/oceanographers/ecologists and anyone else wanting to overlay data, or clip data to, various defined marine regions
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
Not that I know of
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [x] This package does not violate the Terms of Service of any service it interacts with.
    * [x] The repository has continuous integration with Travis CI and/or another service
    * [x] The package contains a vignette
    * [x] The package contains a reasonably complete README with instructions for installing the development version.
    * [x] The package contains unit tests
    * [x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [x] Are there any package dependencies not on CRAN?  __No__
    * [x] Do you intend for this package to go on CRAN?  __Yes__
    * [x] Does the package have a CRAN accepted license?    __Yes__
    * [x] Did `R CMD check` (or `devtools::check()`) produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in circumstances.
