### 2016-06-16 #53 @noamross
Passes all editors' checks, seems like a good fit 😉

Seeking reviewers...
