### 2016-06-24 #53 @mdsumner
# Comments

The `mregions` package provides facilities to easily obtain data from http://www.marineregions.org/ with helpers for working with the data, and including specialist tools for the geo-spatial forms of data (GeoJSON and shapefiles).

The package is structured well using the modern approach with `devtools` and `roxygen2`, and using all the ROpenSci recommended development tools, including version control, automated unit-testing, continuous integration, README and vignettes.

The package is easy to use and seems to do all the things that it's designed to do. I found it hard to see what the package is for overall, since the readme, vignette and documentation are all quite sparse and there's no overall "why this?" message. That may not matter if this is specifically for users of Marine Regions, but it seems this could have a lot more overall appeal to a lot of users.

I am very interested personally in how this fits in with the broader geo-spatial ecosystem in R, and will be happy to contribute in an ongoing way - unfortunately this is all I can give before the review deadline. (I had planned on pull requests for some suggestions, but travel plans mean I just can't get to those in time.)

Thanks very much!

## Installation

The package can be checked out and built easily in RStudio with devtools. It passes check on Windows and Debian, also passes check on devel on Windows.

## Specific comments on code

- I was a little unclear on why  we want to convert to WKT, but I see that it's for `robis::occurrence`. The reasoning for the conversion could be explained.

- `region_names_search` re-runs `region_names`, which is a bit wasteful, it could accept already-obtained output of region_names as an optional input

- `agrep` is used in `region_names_search` for fuzzy matching, but no control is given to the user over how fuzzy. Is it too fuzzy to get "EEA" from "EEZ"?

```{r}
library(mregions)
rn <- region_names()
library(dplyr)
bind_rows(region_names_search("EEZ")) %>% select(title)
```

- `get_region_names` returns a list, but this might as well be a dataframe (?)

```{r,eval=FALSE}
do.call(rbind, lapply(region_names(), as.data.frame, stringsAsFactors = FALSE))
```

- This item is about `robis` rather than `mregions`, but it's important enough to the example used that I think it needs attention.  `robis::occurrence` expects geometry using only longitude/latitude, but as_wkt does not do anything different with a projected Spatial object. It should probably error or auto-transform, rgdal is in use so that's not difficult, (raster has good heuristics for seeing if the assumption of longlat is ok, in the case that an unprojected Spatial object is given without an explicit proj4string). I'm happy to contribute if needed.

This example results in all(?) possible records being queried.

```{r,eval=FALSE}
#' devtools::install_github("iobis/robis")
library('robis')
shp <- region_shp(name = "Belgian Exclusive Economic Zone")
## a rogue user decides to load a projected map somehow
shp <- sp::spTransform(shp, "+proj=laea +ellps=WGS84 +lon_0=2.5 +lat_0=51.6")
wkt <- as_wkt(shp)
## now we are in trouble...
xx <- occurrence("Abra alba", geometry = wkt)
## Retrieved 2000 records of 33188 (6%)
## ...
```

## Questions

- the makefile install for the vignette, is that because you cannot automatically build the figure snapshots from source code, or because robis is not on CRAN? Should mregions list robis as a Suggests dependency and be listed in Remotes?

## ROpensci criteria

Package name, Function and variable naming, Coding style, Code of Conduct Testing, Package scaffolding,  all good.

### Readme

The readme does not describe what the package is for, beyond "get data from ..."

### Documentation

The documentation is very terse, functions do not have Details or Value slots, a few explicit sentences in each would help a lot

- `as_wkt` documentation says it accepts output of "region_geojson" but it also accepts a shapefile name, or a Spatial object
- I suggest to change address of Marine Regions site to "http://www.marineregions.org/"

There is a vignette, but it's a little terse and some of the pre-amble is about installing and is not relevant afaics (move that to readme?).

- `vignette("mregions_vignette")` is hard to remember, perhaps just call the vignette "mregions"?

- `res` is used as the example object name for a few different kinds of objects in the `mregions_vignette`, so there is a lack of clarity on what the workflow is in the sequence of examples.

- The `mregions_vignette` lists iobis/robis but not how to install it, and the vignette uses it but only in "comment" mode. It's not absolutely obvious that robis is a dependency for the vignette material

### Examples

These are good if a little terse in places.

### Console messages

There aren't many of these, but I have not had time to explore in much detail.
