### 2016-06-24 #53 @noamross
Thank your for the review @mdsumner! @sckott will make updates and post a response.
