### 2016-07-13 ropenscilabs/mregions#13 @sckott
review: add more justification/motivation for why to use the pkg
in readme/vignette/pkg level man file
