### 2016-07-13 ropenscilabs/mregions#17 @sckott
review: change region_names output to a data.frame
and make sure all examples work with new output
