### 2016-07-13 ropenscilabs/mregions#18 @sckott
review: change all examples using robis to use rgbif 
or other pkg on CRAN, robis is not on cran so not good fit for this
