### 2016-07-13 ropenscilabs/mregions#19 @sckott
review: function man files improvements
add what they return and make sure param def's are accurate, add `@details` for more details about each fxn
