### 2016-07-13 ropenscilabs/mregions#21 @sckott
review: change address of Marine Regions site to "http://www.marineregions.org/"
in docs
