### 2016-07-13 #53 @sckott
responses to @mdsumner ---> links to issues opened in `ropenscilabs/mregions`

> I found it hard to see what the package is for overall, since the readme, vignette and documentation are all quite sparse and there's no overall "why this?" message. That may not matter if this is specifically for users of Marine Regions, but it seems there are a lot more potential users.

Good point, i'll add some more why you should care bits to various places in the pkg docs (ropenscilabs/mregions#13)

> I am very interested personally in how this fits in with the broader geo-spatial ecosystem in R, and will be happy to contribute in an ongoing way - unfortunately this is all I can give before the review deadline.

Looking forward to any contributions!

> I was a little unclear on why we want to convert to WKT, but I see that it's for `robis::occurrence`. The reasoning for the conversion could be explained.

Right, and WKT is a common geospatial data format, often used to store spatial data in databases, and a number of species occurrence databases expect WKT, including OBIS and the biggest one http://www.gbif.org/ - I'll include more reasoning for this (ropenscilabs/mregions#14)

> `region_names_search` re-runs region_names, which is a bit wasteful, it could accept already-obtained output of `region_names` as an optional input

Good idea, can have it accept either a string or the output of `region_names()` (ropenscilabs/mregions#15)

> `agrep` is used in `region_names_search` for fuzzy matching, but no control is given to the user over how fuzzy. Is it too fuzzy to get "EEA" from "EEZ"?

I can add `...` to the fxn (ropenscilabs/mregions#16)

> `region_names` returns a list, but this might as well be a dataframe (?)

true, i'll do that (ropenscilabs/mregions#17)

> `robis::occurrence` expects geometry using only longitude/latitude, but as_wkt does not do anything different with a projected Spatial object. ...

Please do contribute if you have time, if not, let me know and I'll try

> the makefile install for the vignette, is that because you cannot automatically build the figure snapshots from source code, or because robis is not on CRAN? Should mregions list robis as a Suggests dependency and be listed in Remotes? (I may be unclear on the scope of review here, and what's appropriate for ROpenSci versus CRAN).

it's probably controversial but I don't like vignettes building on package installation, no other languages do this, so i pre-render the vignette to markdown

I just rembered that `robis` is not on CRAN, and it's out of my control, so not sure when it will be there. I'll use a different package that is on CRAN for examples/vignette

(ropenscilabs/mregions#18)

> The readme does not describe what the package is for, beyond "get data from ..."

Good point, will add more info to readme about what it's for (ropenscilabs/mregions#13)

> The documentation is very terse, functions do not have Details or Value slots, a few explicit sentences in each would help a lot

Thanks, will make improvements (ropenscilabs/mregions#19)

> `as_wkt` documentation says it accepts output of "region_geojson" but it also accepts a shapefile name, or a Spatial object

Good catch, i'll fix that (ropenscilabs/mregions#20)

> I suggest to change address of Marine Regions site to "http://www.marineregions.org/"

Will do (ropenscilabs/mregions#21)

> There is a vignette, but it's a little terse and some of the pre-amble is about installing and is not relevant afaics (move that to readme?).

I'll improve the vignette  (ropenscilabs/mregions#22)

> vignette("mregions_vignette") is hard to remember, perhaps just call the vignette "mregions"?

good idea  (ropenscilabs/mregions#23)

> res is used as the example object name for a few different kinds of objects in the mregions_vignette, so there is a lack of clarity on what the workflow is in the sequence of examples.

not sure what you mean, but I'll use different object names to avoid any confusion  (ropenscilabs/mregions#24)

> The mregions_vignette lists iobis/robis but not how to install it, and the vignette uses it but only in "comment" mode. It's not absolutely obvious that robis is a dependency for the vignette material

I just rembered that `robis` is not on CRAN, and it's out of my control, so not sure when it will be there. I'll use a different package that is on CRAN for examples/vignette

> Examples - These are good if a little terse in places

Good point, I'll add more examples  (ropenscilabs/mregions#25)
