### 2016-07-26 #53 @sckott
All issues mentioned above in `ropenscilabs/mregions` have been addressed and closed

Anything else @mdsumner @noamross  ?
