### 2016-07-27 #53 @noamross
Just a few comments in addition to @mdsumner's nice review.  @mdsumner, let us know if @sckott has addressed your concerns to satisfaction.

---

-   You're probably aware your tests are currently failing on Travis, with

         Error: conversion to wkt results in longlat data (@test-as_wkt_projection.R#8)

    I don't get this locally

-   There's also a WARNING that I assume is due to the vignette not being prepared for CRAN yet.
-   Is there a reason we don't have test coverage in CI?
-   Super nitpicky, mostly to try something out: I ran Gabor's new **goodpractice** package on this, and in addition to some complaints about test coverage and long lines, it found a usage of `sapply()` in `rev_geo_code()`.  It's a bit of an odd call, as the output is discarded because this is actually applying a type-checking assertion.  I guess there's no reason for `sapply()` as output is discarded. Switch to `lapply()`?
-  This is just a suggestion, but as there are a variety of geocoding and geoprocessing packages, and their tasks are similar, might it be better to have a function naming scheme that starts with `mr_`?  This way its clear in one's code that one is querying for _marine_ information.
