### 2016-07-28 #53 @sckott
nice, thanks @noamross

> You're probably aware your tests are currently failing on Travis, with

I'll look into that, thanks

> There's also a WARNING that I assume is due to the vignette not being prepared for CRAN yet.

right, i'm aware of that, will remove until on cran

> Is there a reason we don't have test coverage in CI?

don't know why either, can add that.

> sapply() in rev_geo_code() ... Switch to lapply()?

yeah, good thinking!

> might it be better to have a function naming scheme that starts with mr_? This way its clear in one's code that one is querying for marine information.

i think so, so ?:

* `as_wkt` -> `mr_as_wkt`
* `geo_code` -> `mr_geo_code`
* `obis_eez_id` -> `mr_obis_eez_id`
* `place_relations` -> `mr_place_relations`
* `place_types` -> `mr_place_types`
* `records_by_type` -> `mr_records_by_type`
* `region_geojson` -> `mr_region_geojson`
* `region_names` -> `mr_region_names`
* `region_names_search` -> `mr_region_names_search`
* `region_shp` -> `mr_region_shp`
* `rev_geo_code` -> `mr_rev_geo_code`
