### 2016-08-01 #53 @sckott
pretty much ready to go except there's a weird problem in the test suite only on CI seemingly related to`sp`, but not quite sure
