### 2016-08-02 #53 @mdsumner
Oops that's mine, I'll have a look.

ast 13 lines of output:
  3: .ensureIsLonlat(x)
  4: sp::spTransform(x, "+init=EPSG:4326")
  5: sp::spTransform(x, "+init=EPSG:4326")
  6: spTransform(x, CRS(CRSobj), ...)
  7: CRS(CRSobj)
  8: stop(res[[2]])

  testthat results
================================================================
  OK: 0 SKIPPED: 11 FAILED: 1
  1. Error: conversion to wkt results in longlat data
(@test-as_wkt_projection.R#7)

On Tue, 2 Aug 2016 at 08:55 Scott Chamberlain <notifications@github.com>
wrote:

> pretty much ready to go except there's a weird problem in the test suite
> only on CI seemingly related tosp, but not quite sure
>
