### 2016-08-02 #53 @mdsumner
Hi, I'm satisfied that the issues from my review have been addressed. Good job!

Cheers, Mike
