### 2016-08-02 #53 @sckott
cool, now to go to CRAN
